Remove redundant casting to f64

+++ src/views/text_input.rs
@@ -348,8 +348,7 @@ impl TextInput {
         let selection_start_x = selection_start_x.max(node_layout.location.x as f64 - left_padding);
 
         let selection_end_x = virtual_text.hit_position(selection.end).point.x
-            + left_padding as f64
-            - self.clip_start_x;
+            + left_padding - self.clip_start_x;
         let selection_end_x =
             selection_end_x.min(selection_start_x + self.width as f64 + left_padding);
